+++ Microsoft.Azure.Cosmos/src/Query/Core/CoordinatorDistributionPlan/ClientQL/ClientQLMDateTimeLiteral.cs
@@ -6,12 +6,12 @@ namespace Microsoft.Azure.Cosmos.Query.Core.CoordinatorDistributionPlan.ClientQL
 {
     internal class ClientQLMDateTimeLiteral : ClientQLLiteral
     {
-        public ClientQLMDateTimeLiteral(int value)
+        public ClientQLMDateTimeLiteral(long value)
             : base(ClientQLLiteralKind.MDateTime)
         {
             this.Value = value;
         }
 
-        public int Value { get; }
+        public long Value { get; }
     }
 }
