+++ Microsoft.Azure.Cosmos/src/Query/Core/CoordinatorDistributionPlan/ClientQL/ClientQLTupleItemRefScalarExpression.cs
@@ -6,7 +6,7 @@ namespace Microsoft.Azure.Cosmos.Query.Core.CoordinatorDistributionPlan.ClientQL
 {
     internal class ClientQLTupleItemRefScalarExpression : ClientQLScalarExpression
     {
-        public ClientQLTupleItemRefScalarExpression(ClientQLScalarExpression expression, int index) 
+        public ClientQLTupleItemRefScalarExpression(ClientQLScalarExpression expression, long index) 
             : base(ClientQLScalarExpressionKind.TupleItemRef)
         {
             this.Expression = expression;
@@ -15,7 +15,7 @@ namespace Microsoft.Azure.Cosmos.Query.Core.CoordinatorDistributionPlan.ClientQL
 
         public ClientQLScalarExpression Expression { get; }
         
-        public int Index { get; }
+        public long Index { get; }
     }
 
 }
